analysis split into two packages
